pyproject.toml
[project]
name = "instru-sense"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pyyaml",
    "numpy>=2.2.6",
    "python-multipart>=0.0.21",
    "torchcodec>=0.9.1",
    "soundfile>=0.13.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=9.0.2",
    "pytest-asyncio",
    "torch",
    "torchaudio",
]

[tool.pytest.ini_options]
pythonpath = ["."]
